+++ clean_business_chart/tests/clean_business_chart/test_columnchartwithwaterfall.py
@@ -261,12 +261,23 @@ def test__dataframe_full_year():
     message  = "Test 1 - ColumnWithWaterfall._dataframe_full_year returned {0} instead of {1}".format(actual, expected)
     assert actual == expected, message
 
-    # Test 2 - only dataframe supported
+    # Test 2 - good empty dataframe with year and month columns
+    dataset = pd.DataFrame({'Year' : [], 
+                            'Month': [],
+                            'AC'   : []})
+    testvar  = ColumnWithWaterfall(test=True)
+    expected = {'Year': {}, 'Month': {}, 'AC': {}}
+    actual   = testvar._dataframe_full_year(dataset)
+    actual   = actual.to_dict()
+    message  = "Test 2 - ColumnWithWaterfall._dataframe_full_year returned {0} instead of {1}".format(actual, expected)
+    assert actual == expected, message
+
+    # Test 3 - only dataframe supported
     with pytest.raises(ValueError):
         testvar = ColumnWithWaterfall(test=True)
         testvar._dataframe_full_year("This is a string")
 
-    # Test 3 - a dataframe with more than one year
+    # Test 4 - a dataframe with more than one year
     with pytest.raises(ValueError):
         dataset = pd.DataFrame({'Year' : ['2022', '2021'], 
                                 'Month': ['01', '02'],
@@ -274,7 +285,7 @@ def test__dataframe_full_year():
         testvar  = ColumnWithWaterfall(test=True)
         actual   = testvar._dataframe_full_year(dataset)
 
-    # Test 4 - a dataframe with missing column
+    # Test 5 - a dataframe with missing column
     with pytest.raises(ValueError):
         dataset = pd.DataFrame({'Year' : ['2022', '2022'], 
                                 'AC': [35, 33]})
@@ -341,7 +352,7 @@ def test__dataframe_translate_field_headers():
 
 
 def test__convert_data_list_of_lists_to_dataframe():
-    # Test 1 - good dataframe with string year values and integer and float values
+    # Test 1 - good list of lists 
     dataset = [['Year', 'Month', 'AC', 'PL', 'FC'], [2022, 1, 35, 33, 0], [2022, 2, 38, 40, 0], [2022, 3, 29, 35, 0]]
     testvar  = ColumnWithWaterfall(test=True)
     expected =  {'Year': {0: 2022, 1: 2022, 2: 2022}, 
@@ -364,3 +375,108 @@ def test__convert_data_list_of_lists_to_dataframe():
         dataset = [['Year', 'Month', 'AC', 'PL', 'FC'], [2022, 1, 35, 33, 0], "This element is a string", [2022, 3, 29, 35, 0]]
         testvar = ColumnWithWaterfall(test=True)
         testvar._convert_data_list_of_lists_to_dataframe(dataset)
+
+
+def test__dataframe_handle_previous_year():
+    # Test 1 - good dataframe with two years of information and actual of previous year matches previous year of highest year
+    dataset = pd.DataFrame({'Year' : ['2021', '2021', '2021', '2021']+['2022', '2022', '2022', '2022'], 
+                            'Month': ['02', '04', '07', '08']+['02', '04', '07', '08'],
+                            'PY'   : [32, 38, 12, 29]+[35, 33, 17, 41], 
+                            'AC'   : [35, 33, 17, 41]+[37, 36, 25, 38]})
+    testvar  = ColumnWithWaterfall(test=True)
+    expected =  {'Year': {0: '2022', 1: '2022', 2: '2022', 3: '2022', 4: '2022', 5: '2022', 6: '2022', 7: '2022', 8: '2022', 9: '2022', 10: '2022', 11: '2022'}, 
+                 'Month': {0: '01', 1: '02', 2: '03', 3: '04', 4: '05', 5: '06', 6: '07', 7: '08', 8: '09', 9: '10', 10: '11', 11: '12'}, 
+                 'PY': {0: 0.0, 1: 35.0, 2: 0.0, 3: 33.0, 4: 0.0, 5: 0.0, 6: 17.0, 7: 41.0, 8: 0.0, 9: 0.0, 10: 0.0, 11: 0.0}, 
+                 'AC': {0: 0.0, 1: 37.0, 2: 0.0, 3: 36.0, 4: 0.0, 5: 0.0, 6: 25.0, 7: 38.0, 8: 0.0, 9: 0.0, 10: 0.0, 11: 0.0}}
+    actual   = testvar._dataframe_handle_previous_year(dataset)
+    actual   = actual.to_dict()
+    message  = "Test 1 - ColumnWithWaterfall._dataframe_handle_previous_year returned {0} instead of {1}".format(actual, expected)
+    assert actual == expected, message
+
+    # Test 2 - good dataframe with one year of information
+    dataset = pd.DataFrame({'Year' : ['2022', '2022', '2022', '2022'], 
+                            'Month': ['02', '04', '07', '08'],
+                            'PY'   : [32, 38, 12, 29], 
+                            'AC'   : [35, 33, 17, 41]})
+    testvar  = ColumnWithWaterfall(test=True)
+    expected = {'Year': {0: '2022', 1: '2022', 2: '2022', 3: '2022', 4: '2022', 5: '2022', 6: '2022', 7: '2022', 8: '2022', 9: '2022', 10: '2022', 11: '2022'}, 
+                'Month': {0: '01', 1: '02', 2: '03', 3: '04', 4: '05', 5: '06', 6: '07', 7: '08', 8: '09', 9: '10', 10: '11', 11: '12'}, 
+                'PY': {0: 0.0, 1: 32.0, 2: 0.0, 3: 38.0, 4: 0.0, 5: 0.0, 6: 12.0, 7: 29.0, 8: 0.0, 9: 0.0, 10: 0.0, 11: 0.0}, 
+                'AC': {0: 0.0, 1: 35.0, 2: 0.0, 3: 33.0, 4: 0.0, 5: 0.0, 6: 17.0, 7: 41.0, 8: 0.0, 9: 0.0, 10: 0.0, 11: 0.0}}
+    actual   = testvar._dataframe_handle_previous_year(dataset)
+    actual   = actual.to_dict()
+    message  = "Test 2 - ColumnWithWaterfall._dataframe_handle_previous_year returned {0} instead of {1}".format(actual, expected)
+    assert actual == expected, message
+
+    # Test 3 - good dataframe with two years of information and actual of previous year does not match previous year of highest year
+    with pytest.raises(ValueError):
+        dataset = pd.DataFrame({'Year' : ['2021', '2021', '2021', '2021']+['2022', '2022', '2022', '2022'], 
+                                'Month': ['02', '04', '07', '08']+['02', '04', '07', '08'],
+                                'PY'   : [32, 38, 12, 29]+[135, 233, 317, 441], 
+                                'AC'   : [35, 33, 17, 41]+[37, 36, 25, 38]})
+        testvar = ColumnWithWaterfall(test=True)
+        testvar._dataframe_handle_previous_year(dataset)
+
+    # Test 4 - good dataframe missing the year column
+    with pytest.raises(ValueError):
+        dataset = pd.DataFrame({'Month': ['02', '04', '07', '08'],
+                                'PY'   : [32, 38, 12, 29], 
+                                'AC'   : [35, 33, 17, 41]})
+        testvar = ColumnWithWaterfall(test=True)
+        testvar._dataframe_handle_previous_year(dataset)
+
+    # Test 5 - only dataframe supported
+    with pytest.raises(ValueError):
+        testvar = ColumnWithWaterfall(test=True)
+        testvar._dataframe_handle_previous_year("This is a string")
+
+
+def test__dataframe_to_dictionary():
+    # Test 1 - good dataframe with one complete years of information
+    dataset = pd.DataFrame({'Year' : ['2022', '2022', '2022', '2022', '2022', '2022', '2022', '2022', '2022', '2022', '2022', '2022'], 
+                            'Month': ['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12'],
+                            'PY'   : [35, 33, 17, 41, 25, 32, 28, 34, 27, 24, 37, 36], 
+                            'AC'   : [37, 36, 25, 38, 28, 27, 34, 33, 29, 26, 38, 39]})
+    testvar  = ColumnWithWaterfall(test=True)
+    expected =  {'Year': '2022', 
+                 'PY': [35, 33, 17, 41, 25, 32, 28, 34, 27, 24, 37, 36], 
+                 'AC': [37, 36, 25, 38, 28, 27, 34, 33, 29, 26, 38, 39]}
+    actual   = testvar._dataframe_to_dictionary(dataset)
+    message  = "Test 1 - ColumnWithWaterfall._dataframe_to_dictionary returned {0} instead of {1}".format(actual, expected)
+    assert actual == expected, message
+
+    # Test 2 - good dataframe with four months of one year of information
+    dataset = pd.DataFrame({'Year' : ['2022', '2022', '2022', '2022'], 
+                            'Month': ['02', '04', '07', '08'],
+                            'PY'   : [32, 38, 12, 29], 
+                            'AC'   : [35, 33, 17, 41]})
+    testvar  = ColumnWithWaterfall(test=True)
+    expected =  {'Year': '2022', 'PY': [32, 38, 12, 29], 'AC': [35, 33, 17, 41]}
+    actual   = testvar._dataframe_to_dictionary(dataset)
+    message  = "Test 2 - ColumnWithWaterfall._dataframe_to_dictionary returned {0} instead of {1}".format(actual, expected)
+    assert actual == expected, message
+
+    # Test 3 - good dataframe with four months of one year of information and with string values in scenarios
+    dataset = pd.DataFrame({'Year' : ['2022', '2022', '2022', '2022'], 
+                            'Month': ['02', '04', '07', '08'],
+                            'PY'   : [32, '38', 12, 29], 
+                            'AC'   : [35, 33, '17', 41]})
+    testvar  = ColumnWithWaterfall(test=True)
+    expected =  {'Year': '2022', 'PY': [32, 38, 12, 29], 'AC': [35, 33, 17, 41]}
+    actual   = testvar._dataframe_to_dictionary(dataset)
+    message  = "Test 3 - ColumnWithWaterfall._dataframe_to_dictionary returned {0} instead of {1}".format(actual, expected)
+    assert actual == expected, message
+
+
+    # Test 4 - good dataframe missing the year column
+    with pytest.raises(ValueError):
+        dataset = pd.DataFrame({'Month': ['02', '04', '07', '08'],
+                                'PY'   : [32, 38, 12, 29], 
+                                'AC'   : [35, 33, 17, 41]})
+        testvar = ColumnWithWaterfall(test=True)
+        testvar._dataframe_to_dictionary(dataset)
+
+    # Test 5 - only dataframe supported
+    with pytest.raises(ValueError):
+        testvar = ColumnWithWaterfall(test=True)
+        testvar._dataframe_to_dictionary("This is a string")